chore: Update .gitignore and myacg_manager.py
This commit updates the .gitignore file to remove the 'save/*' entry and adds the 'save' directory to be ignored. It also modifies the myacg_manager.py file to import the NoSuchElementException from the selenium.common.exceptions module. Additionally, it adds code to check if the order is using a coupon and locates the order establishment date.